pyproject.toml
[project]
name = "kronos"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy",
    "pandas==2.2.2",
    "torch",
    "einops==0.8.1",
    "huggingface_hub==0.33.1",
    "matplotlib==3.9.3",
    "tqdm==4.67.1",
    "safetensors>=0.6.2",
    "flask==2.3.3",
    "flask-cors==4.0.0",
    "plotly==5.17.0",
    "requests>=2.32.5",
    "polygon-api-client>=1.15.3",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
    "dotenv",
    "pyqlib"
]
